Resolved decord-caused errors in training and env install.
Made changes in `conda_env.yaml` and `train.py` that solves issues related to decord. In the first one, `decord` was not a requirement and thus installation of an env was not complete without installing it. For the second one, importing `decord` before `torch` sometimes caused a segfault.

This PR changes them both.